Hago un nuevo componente para manejar las notificaciones
