+++ backend/src/main/java/app/model/notificador/Notificacion.java
@@ -28,12 +28,12 @@ public class Notificacion {
       this.leida = false;
   }
 
-    public Notificacion(Mensaje mensaje, Perfil perfil, String link) {
-        this.mensaje = mensaje;
-        this.perfil = perfil;
-        this.link = link;
-        this.leida = false;
-    }
+  public Notificacion(Mensaje mensaje, Perfil perfil, String link) {
+      this.mensaje = mensaje;
+      this.perfil = perfil;
+      this.link = link;
+      this.leida = false;
+  }
 
   public void marcarLeida() {
       this.leida = true;
